Split UI rendering files by concern
